+++ app/Http/Controllers/NewsCategoryController.php
@@ -2,12 +2,18 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Section;
 use App\NewsCategory;
 use Illuminate\Http\Request;
 
 class NewsCategoryController extends Controller
 {
+    public $section;
 
+    public function __construct()
+    {
+        $this->section = Section::find(config('app.enabled_sections.noticias'));
+    }
 
     public function index()
     {
@@ -16,6 +22,7 @@ class NewsCategoryController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('write', $this->section);
         $category = NewsCategory::create([]);
         foreach ($request->languages as $language) {
             $category->languages()->create($language);
@@ -25,6 +32,7 @@ class NewsCategoryController extends Controller
 
     public function update(Request $request, $id)
     {
+        $this->authorize('write', $this->section);
         $category = NewsCategory::findOrFail($id);
         foreach ($request->languages as $language) {
             $category->lang($language["language_id"])->update($language);
@@ -34,6 +42,7 @@ class NewsCategoryController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('delete', $this->section);
         $category = NewsCategory::findOrFail($id);
         $category->languages()->delete();
         $category->delete();
